mvMetricsWindow.h, mvInputItems.h, mvDragItems.h splited into .h and .cpp files, mvDragInt3 - empty constructor, cmake file updated with mvMetricsWindow.cpp, mvInputItems.cpp and mvDragItems.cpp
